+++ src/ufbget.f
@@ -109,7 +109,7 @@ C  ---------------------------------------------------------
          GOTO 20
       ELSEIF(ITP(NODE).EQ.1) THEN
          CALL UPB8(IVAL,NBIT(N),MBIT(N),MBAY(1,LUN))
-         NBMP=IVAL; CALL USRTPL(LUN,N,NBMP)
+         NBMP=INT(IVAL); CALL USRTPL(LUN,N,NBMP)
          GOTO 10
       ENDIF
       ENDDO
